add a custom component Searchbox and a custom hook useFuzzySearch, and use them in both index and map pages
